Added camshift detector

+++ src/target_tracking/include/target_tracking/rransac/open_cv_access.h
@@ -1,12 +1,12 @@
     #pragma once
 
-    class OpenCVPointAccess : public rransac::BaseAccessType<cv::Point2d>
+    class OpenCVPointAccess : public rransac::BaseAccessType<cv::Point2f>
     {
     public:
         OpenCVPointAccess() {};
         ~OpenCVPointAccess() {};
 
     private:
-        const double x(const cv::Point2d& p) { return p.x; }
-        const double y(const cv::Point2d& p) { return p.y; }
+        const double x(const cv::Point2f& p) { return (double)p.x; }
+        const double y(const cv::Point2f& p) { return (double)p.y; }
     };
